+++ src/main/kotlin/dev/codeswamp/domain/article/infrastructure/graph/node/ArticleNode.kt
@@ -1,4 +1,4 @@
-package dev.codeswamp.domain.article.infrastructure.persistence.entity.node
+package dev.codeswamp.domain.article.infrastructure.graph.node
 
 import org.springframework.data.neo4j.core.schema.GeneratedValue
 import org.springframework.data.neo4j.core.schema.Id
