+++ src/main/kotlin/dev/codeswamp/domain/article/infrastructure/persistence/repository/ArticleViewJpaRepository.kt
@@ -1,0 +1,7 @@
+package dev.codeswamp.domain.article.infrastructure.persistence.repository
+
+import dev.codeswamp.domain.article.infrastructure.persistence.entity.ArticleView
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface ArticleViewJpaRepository : JpaRepository<ArticleView, Long> {
+}
